# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * KELIN MD WhatsApp Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BotStatusStatus = typeof BotStatusStatus[keyof typeof BotStatusStatus];


export const BotStatusStatus = {
  connected: 'connected',
  disconnected: 'disconnected',
  connecting: 'connecting',
  pairing: 'pairing',
} as const;

export interface BotStatus {
  connected: boolean;
  status: BotStatusStatus;
  /** Uptime in seconds */
  uptime: number;
  ramUsageMb: number;
  ramTotalMb: number;
  cpuUsage: number;
  totalUsers: number;
  totalGroups: number;
  totalCommands: number;
  pluginCount: number;
  /** @nullable */
  connectedNumber: string | null;
  botName: string;
  version: string;
  /** @nullable */
  pairingCode?: string | null;
}

export interface PairingRequest {
  /** Phone number with country code (e.g. +1234567890) */
  phoneNumber: string;
}

export interface PairingResponse {
  pairingCode: string;
  phoneNumber: string;
}

export interface SessionInfo {
  connected: boolean;
  hasSession: boolean;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  name?: string | null;
  sessionPath?: string;
  /** @nullable */
  connectedAt?: string | null;
}

export type PluginCategory = typeof PluginCategory[keyof typeof PluginCategory];


export const PluginCategory = {
  main: 'main',
  group: 'group',
  admin: 'admin',
  owner: 'owner',
  download: 'download',
  ai: 'ai',
  anime: 'anime',
  fun: 'fun',
  games: 'games',
  media: 'media',
  search: 'search',
  utilities: 'utilities',
} as const;

export interface Plugin {
  id: string;
  name: string;
  category: PluginCategory;
  enabled: boolean;
  commandCount: number;
  description: string;
  version?: string;
}

export interface PluginToggle {
  enabled: boolean;
}

export interface Command {
  name: string;
  description: string;
  category: string;
  usage: string;
  aliases?: string[];
  cooldown: number;
  isOwner: boolean;
  isAdmin: boolean;
  isPremium: boolean;
}

export type LogEntryLevel = typeof LogEntryLevel[keyof typeof LogEntryLevel];


export const LogEntryLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  debug: 'debug',
} as const;

export interface LogEntry {
  id: string;
  timestamp: string;
  level: LogEntryLevel;
  message: string;
  /** @nullable */
  details?: string | null;
}

export type FileEntryType = typeof FileEntryType[keyof typeof FileEntryType];


export const FileEntryType = {
  file: 'file',
  directory: 'directory',
} as const;

export interface FileEntry {
  name: string;
  path: string;
  type: FileEntryType;
  size: number;
  modified: string;
}

export interface FileContent {
  path: string;
  content: string;
  size: number;
}

export interface BotSettings {
  ownerName: string;
  ownerNumber: string;
  botName: string;
  prefix: string;
  themeEmoji: string;
  footer: string;
  mongoUri: string;
  sessionFolder: string;
  premiumUsers: string[];
  adminNumbers: string[];
  autoRead?: boolean;
  autoReact?: boolean;
  antiLink?: boolean;
  antiSpam?: boolean;
  antiDelete?: boolean;
  welcomeMessage?: boolean;
  xpSystem?: boolean;
}

export interface BotSettingsInput {
  ownerName?: string;
  ownerNumber?: string;
  botName?: string;
  prefix?: string;
  themeEmoji?: string;
  footer?: string;
  mongoUri?: string;
  sessionFolder?: string;
  premiumUsers?: string[];
  adminNumbers?: string[];
  autoRead?: boolean;
  autoReact?: boolean;
  antiLink?: boolean;
  antiSpam?: boolean;
  antiDelete?: boolean;
  welcomeMessage?: boolean;
  xpSystem?: boolean;
}

export interface BackupInfo {
  id: string;
  filename: string;
  createdAt: string;
  sizeMb: number;
}

export interface RestoreRequest {
  backupId: string;
}

export interface ApiKeyStatus {
  name: string;
  configured: boolean;
  /** @nullable */
  working: boolean | null;
  description?: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetLogsParams = {
level?: GetLogsLevel;
limit?: number;
};

export type GetLogsLevel = typeof GetLogsLevel[keyof typeof GetLogsLevel];


export const GetLogsLevel = {
  all: 'all',
  info: 'info',
  warn: 'warn',
  error: 'error',
  debug: 'debug',
} as const;

export type ListFilesParams = {
path?: string;
};

export type GetFileContentParams = {
path: string;
};