# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KELIN MD WhatsApp Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/pair:
    post:
      operationId: requestPairing
      tags: [bot]
      summary: Request a pairing code
      description: Generates a WhatsApp pairing code for the given phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairingRequest"
      responses:
        "200":
          description: Pairing code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /plugins/{pluginId}/toggle:
    patch:
      operationId: togglePlugin
      tags: [plugins]
      summary: Enable or disable a plugin
      parameters:
        - name: pluginId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginToggle"
      responses:
        "200":
          description: Plugin toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plugin"
        "404":
          description: Plugin not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in a directory
      parameters:
        - name: path
          in: query
          schema:
            type: string
            default: "."
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileEntry"
  /files/content:
    get:
      operationId: getFileContent
      tags: [files]
      summary: Get file content
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /restore:
    post:
      operationId: restoreBackup
      tags: [backup]
      summary: Restore from backup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreRequest"
      responses:
        "200":
          description: Restore initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Backup not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    PairingRequest:
      type: object
      required: [phoneNumber]
      properties:
        phoneNumber:
          type: string
          description: Phone number with country code (e.g. +1234567890)
    PairingResponse:
      type: object
      required: [pairingCode, phoneNumber]
      properties:
        pairingCode:
          type: string
        phoneNumber:
          type: string
...
    Plugin:
      type: object
      required: [id, name, category, enabled, commandCount, description]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [main, group, admin, owner, download, ai, anime, fun, games, media, search, utilities]
        enabled:
          type: boolean
        commandCount:
          type: integer
        description:
          type: string
        version:
          type: string
    PluginToggle:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
...
    FileEntry:
      type: object
      required: [name, path, type, size, modified]
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        size:
          type: integer
        modified:
          type: string
    FileContent:
      type: object
      required: [path, content, size]
      properties:
        path:
          type: string
        content:
          type: string
        size:
          type: integer
...
    RestoreRequest:
      type: object
      required: [backupId]
      properties:
        backupId:
          type: string
...
    SuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string