# lib/api-zod/src/generated/types/fileContent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * KELIN MD WhatsApp Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface FileContent {
  path: string;
  content: string;
  size: number;
}